+++ projects/project-4/src/thread_pool/naive_thread_pool.rs
@@ -6,7 +6,7 @@ pub struct NaiveThreadPool {
 }
 
 impl ThreadPool for NaiveThreadPool {
-    fn new(_threads: u32) -> Result<Self> {
+    fn new(_threads: usize) -> Result<Self> {
         Ok(NaiveThreadPool {})
     }
